server/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.5.7
Django==4.2.2
django-clearcache==1.2.1
django-cors-headers==4.0.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-gis==1.0
GDAL==3.7.1.1
gunicorn==20.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
